refactor: checkbox src
