+++ packages/checkbox/src/checkbox-context.ts
@@ -1,0 +1,8 @@
+import { createContext } from "@chakra-ui/react-utils"
+import { CheckboxGroupContext } from "./checkbox-types"
+
+export const [CheckboxGroupProvider, useCheckboxGroupContext] =
+  createContext<CheckboxGroupContext>({
+    name: "CheckboxGroupContext",
+    strict: false,
+  })
