+++ packages/checkbox/src/checkbox.tsx
@@ -4,16 +4,16 @@ import {
   HTMLChakraProps,
   omitThemingProps,
   PropsOf,
-  SystemProps,
   SystemStyleObject,
   ThemingProps,
   useMultiStyleConfig,
 } from "@chakra-ui/system"
 import { callAll, cx, Omit, __DEV__ } from "@chakra-ui/utils"
-import * as React from "react"
-import { useCheckboxGroupContext } from "./checkbox-group"
+import { cloneElement, useMemo } from "react"
+import { useCheckboxGroupContext } from "./checkbox-context"
 import { CheckboxIcon } from "./checkbox-icon"
-import { useCheckbox, UseCheckboxProps } from "./use-checkbox"
+import { CheckboxOptions, UseCheckboxProps } from "./checkbox-types"
+import { useCheckbox } from "./use-checkbox"
 
 const CheckboxControl = chakra("span", {
   baseStyle: {
@@ -26,7 +26,7 @@ const CheckboxControl = chakra("span", {
   },
 })
 
-const Label = chakra("label", {
+const CheckboxRoot = chakra("label", {
   baseStyle: {
     cursor: "pointer",
     display: "inline-flex",
@@ -47,33 +47,8 @@ export interface CheckboxProps
   extends CheckboxControlProps,
     BaseInputProps,
     ThemingProps<"Checkbox">,
-    UseCheckboxProps {
-  /**
-   * The spacing between the checkbox and its label text
-   * @default 0.5rem
-   * @type SystemProps["marginLeft"]
-   */
-  spacing?: SystemProps["marginLeft"]
-  /**
-   * The color of the checkbox icon when checked or indeterminate
-   */
-  iconColor?: string
-  /**
-   * The size of the checkbox icon when checked or indeterminate
-   */
-  iconSize?: string | number
-  /**
-   * The checked icon to use
-   *
-   * @type React.ReactElement
-   * @default CheckboxIcon
-   */
-  icon?: React.ReactElement
-  /**
-   * Additional props to be forwarded to the `input` element
-   */
-  inputProps?: React.InputHTMLAttributes<HTMLInputElement>
-}
+    UseCheckboxProps,
+    CheckboxOptions {}
 
 /**
  * Checkbox
@@ -83,7 +58,10 @@ export interface CheckboxProps
  *
  * @see Docs https://chakra-ui.com/checkbox
  */
-export const Checkbox = forwardRef<CheckboxProps, "input">((props, ref) => {
+export const Checkbox = forwardRef<CheckboxProps, "input">(function Checkbox(
+  props,
+  ref,
+) {
   const group = useCheckboxGroupContext()
 
   const mergedProps = { ...group, ...props } as CheckboxProps
@@ -128,7 +106,7 @@ export const Checkbox = forwardRef<CheckboxProps, "input">((props, ref) => {
     onChange,
   })
 
-  const iconStyles: SystemStyleObject = React.useMemo(
+  const iconStyles: SystemStyleObject = useMemo(
     () => ({
       opacity: state.isChecked || state.isIndeterminate ? 1 : 0,
       transform:
@@ -140,14 +118,14 @@ export const Checkbox = forwardRef<CheckboxProps, "input">((props, ref) => {
     [iconColor, iconSize, state.isChecked, state.isIndeterminate, styles.icon],
   )
 
-  const clonedIcon = React.cloneElement(icon, {
+  const clonedIcon = cloneElement(icon, {
     __css: iconStyles,
     isIndeterminate: state.isIndeterminate,
     isChecked: state.isChecked,
   })
 
   return (
-    <Label
+    <CheckboxRoot
       __css={styles.container}
       className={cx("chakra-checkbox", className)}
       {...getRootProps()}
@@ -175,7 +153,7 @@ export const Checkbox = forwardRef<CheckboxProps, "input">((props, ref) => {
           {children}
         </chakra.span>
       )}
-    </Label>
+    </CheckboxRoot>
   )
 })
