+++ packages/checkbox/src/index.ts
@@ -1,4 +1,8 @@
-export * from "./checkbox-group"
-export * from "./checkbox"
-export * from "./use-checkbox"
-export * from "./use-checkbox-group"
+export { CheckboxGroup, type CheckboxGroupProps } from "./checkbox-group"
+export { Checkbox, type CheckboxProps } from "./checkbox"
+export { useCheckbox, type UseCheckboxReturn } from "./use-checkbox"
+export type { UseCheckboxGroupProps, UseCheckboxProps } from "./checkbox-types"
+export {
+  useCheckboxGroup,
+  type UseCheckboxGroupReturn,
+} from "./use-checkbox-group"
