+++ packages/checkbox/src/use-checkbox.ts
@@ -9,106 +9,9 @@ import {
 import { mergeRefs, PropGetter } from "@chakra-ui/react-utils"
 import { callAllHandlers, dataAttr, focus, omit } from "@chakra-ui/utils"
 import { visuallyHiddenStyle } from "@chakra-ui/visually-hidden"
-import React, {
-  ChangeEvent,
-  KeyboardEvent,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react"
 import { trackFocusVisible } from "@zag-js/focus-visible"
-
-export interface UseCheckboxProps {
-  /**
-   * If `true`, the checkbox will be checked.
-   * You'll need to pass `onChange` to update its value (since it is now controlled)
-   */
-  isChecked?: boolean
-  /**
-   * If `true`, the checkbox will be indeterminate.
-   * This only affects the icon shown inside checkbox
-   * and does not modify the isChecked property.
-   */
-  isIndeterminate?: boolean
-  /**
-   * If `true`, the checkbox will be disabled
-   */
-  isDisabled?: boolean
-  /**
-   * If `true` and `isDisabled` is passed, the checkbox will
-   * remain tabbable but not interactive
-   */
-  isFocusable?: boolean
-  /**
-   * If `true`, the checkbox will be readonly
-   */
-  isReadOnly?: boolean
-  /**
-   * If `true`, the checkbox is marked as invalid.
-   * Changes style of unchecked state.
-   */
-  isInvalid?: boolean
-  /**
-   * If `true`, the checkbox input is marked as required,
-   * and `required` attribute will be added
-   */
-  isRequired?: boolean
-  /**
-   * If `true`, the checkbox will be initially checked.
-   */
-  defaultChecked?: boolean
-  /**
-   * The callback invoked when the checked state of the `Checkbox` changes.
-   */
-  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void
-  /**
-   * The callback invoked when the checkbox is blurred (loses focus)
-   */
-  onBlur?: (event: React.FocusEvent<HTMLInputElement>) => void
-  /**
-   * The callback invoked when the checkbox is focused
-   */
-  onFocus?: (event: React.FocusEvent<HTMLInputElement>) => void
-  /**
-   * The name of the input field in a checkbox
-   * (Useful for form submission).
-   */
-  name?: string
-  /**
-   * The value to be used in the checkbox input.
-   * This is the value that will be returned on form submission.
-   */
-  value?: string | number
-  /**
-   * id assigned to input
-   */
-  id?: string
-  /**
-   * Defines the string that labels the checkbox element.
-   */
-  "aria-label"?: string
-  /**
-   * Refers to the `id` of the element that labels the checkbox element.
-   */
-  "aria-labelledby"?: string
-  "aria-invalid"?: true | undefined
-  "aria-describedby"?: string
-
-  tabIndex?: number
-}
-
-export interface CheckboxState {
-  isInvalid: boolean | undefined
-  isFocused: boolean
-  isChecked: boolean
-  isActive: boolean
-  isHovered: boolean
-  isIndeterminate: boolean | undefined
-  isDisabled: boolean | undefined
-  isReadOnly: boolean | undefined
-  isRequired: boolean | undefined
-}
+import { useCallback, useEffect, useRef, useState } from "react"
+import { CheckboxState, UseCheckboxProps } from "./checkbox-types"
 
 /**
  * useCheckbox that provides all the state and focus management logic
@@ -179,7 +82,7 @@ export function useCheckbox(props: UseCheckboxProps = {}) {
   )
 
   const handleChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
+    (event: React.ChangeEvent<HTMLInputElement>) => {
       if (isReadOnly || isDisabled) {
         event.preventDefault()
         return
@@ -231,7 +134,7 @@ export function useCheckbox(props: UseCheckboxProps = {}) {
   const trulyDisabled = isDisabled && !isFocusable
 
   const onKeyDown = useCallback(
-    (event: KeyboardEvent) => {
+    (event: React.KeyboardEvent) => {
       if (event.key === " ") {
         setActive.on()
       }
@@ -240,7 +143,7 @@ export function useCheckbox(props: UseCheckboxProps = {}) {
   )
 
   const onKeyUp = useCallback(
-    (event: KeyboardEvent) => {
+    (event: React.KeyboardEvent) => {
       if (event.key === " ") {
         setActive.off()
       }
